requirements.txt
pyautogui
pywin32
opencv-python
paddlepaddle-gpu
paddleocr
requests
setuptools
pynput
